refactor find_relative_frequency
